Renamed trend analysis main code file (modules.py) to have a more obvious name (trend_analysis).
