+++ DataAnalyzer/TrendFinder/trend_analysis.py
